requirements.txt
aiohttp==3.6.2
astroid==2.4.0
async-timeout==3.0.1
attrs==19.3.0
cffi==1.14.0
chardet==3.0.4
discord.py==1.3.3
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.5
mypy==0.770
mypy-extensions==0.4.3
pycparser==2.20
pygobject==3.26.1
pylint==2.5.0
PyNaCl==1.3.0
python-apt==1.6.5+ubuntu0.2
python-dotenv==0.13.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
unattended-upgrades==0.1
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
